Refactor to Server Component and extract task logic
- Convert page.tsx to async Server Component using auth() directly
- Extract Google Tasks API logic into src/lib/tasks.ts
- Extract UI into TaskList and LoginButton components
- Add loading.tsx skeleton UI for Suspense fallback
- Simplify route.ts by delegating to fetchTodayTasks
- Add .claude/settings.local.json to .gitignore
